chore(client): 전반적인 코드 리팩토링
