+++ apps/client/src/components/qna/QnaBox/index.tsx
@@ -1,47 +1,71 @@
 import styled from '@emotion/styled';
+import { IconHeartFill, IconHeartRegular, IconReplyRegular } from '@seed-design/icon';
 import { colors } from '@sickgyun/design-token';
-import { Button, Flex, Stack, Text } from '@sickgyun/ui';
+import { Flex, Stack, Text } from '@sickgyun/ui';
+import QnaCategory from '../QnaCategory';
+import type { Qna } from '@/types/qna';
+import { timeAgo } from '@/utils/timeAgo';
+import { useGetQnaLike } from '@/hooks/api/qna/useGetQnaLike';
 import { useRouter } from 'next/navigation';
-import { useState } from 'react';
-import { QNA_CATEGORY } from '@/constants/qna';
 
-const QnaBox = () => {
+type qnaChildBoxProps = {
+  id: number;
+  category: string;
+  title: string;
+  content: string;
+  createTime: string;
+  likeCount: number;
+  commentCount: number;
+};
+
+const QnaBox = ({
+  id,
+  category,
+  title,
+  content,
+  createTime,
+  likeCount,
+  commentCount,
+}: qnaChildBoxProps) => {
   const router = useRouter();
-  const [activeCategoryIndex, setActiveCategoryIndex] = useState(0);
 
-  const handleGoQnaWritePage = () => {
-    router.push('/qna/write');
+  const handleGoDetailQnaPage = (id: number) => {
+    router.push(`/qna/${id}`);
   };
 
-  const handleGoCategoryQnaPage = (id: number, qnaType: string) => {
-    setActiveCategoryIndex(id);
-    router.push(`/qna?category=${qnaType}`);
-  };
+  const { qnaLike } = useGetQnaLike(Number(id));
 
   return (
-    <StyledQnaBox>
-      <Stack
-        direction="horizontal"
-        align="center"
-        justify="center"
-        style={{ height: '90px', borderBottom: `1px solid ${colors.white}` }}
-      >
-        <Button width="80%" onClick={() => handleGoQnaWritePage()}>
-          질문하기
-        </Button>
-      </Stack>
+    <StyledQnaBox key={id} onClick={() => handleGoDetailQnaPage(id)}>
       <Flex direction="column">
-        {QNA_CATEGORY.map((category) => (
-          <StyledQnaCategory
-            onClick={() => handleGoCategoryQnaPage(category.id, category.qnaType)}
-            isActive={category.id === activeCategoryIndex}
-          >
-            <Stack direction="horizontal" spacing={10}>
-              <Text>{category.emoji}</Text>
-              <Text fontType="body1">{category.qnaTitle}</Text>
+        <StyledQnaContentWrapper>
+          <QnaCategory questionType={category as Qna} />
+          <Text fontType="h4">{title}</Text>
+          <StyledQnaContent>{content}</StyledQnaContent>
+          <Flex justify="space-between" style={{ width: '100%' }}>
+            <Stack spacing={12} direction="horizontal">
+              <Text fontType="body2">{timeAgo(createTime)}</Text>
             </Stack>
-          </StyledQnaCategory>
-        ))}
+            <Stack spacing={12} direction="horizontal">
+              <Stack direction="horizontal" align="center" spacing={3}>
+                {qnaLike ? (
+                  <IconHeartFill width={16} height={16} color={colors.red} />
+                ) : (
+                  <IconHeartRegular width={16} height={16} />
+                )}
+                <Text fontType="body2" style={{ marginTop: '2px' }}>
+                  {likeCount}
+                </Text>
+              </Stack>
+              <Stack direction="horizontal" align="center" spacing={3}>
+                <IconReplyRegular width={16} height={16} color={colors.black} />
+                <Text fontType="body2" style={{ marginTop: '2px' }}>
+                  {commentCount}
+                </Text>
+              </Stack>
+            </Stack>
+          </Flex>
+        </StyledQnaContentWrapper>
       </Flex>
     </StyledQnaBox>
   );
@@ -51,28 +75,27 @@ export default QnaBox;
 
 const StyledQnaBox = styled.div`
   width: 100%;
-  height: 300px;
-  background-color: ${({ theme }) => theme.colors.white};
-  border-radius: 12px;
+  border-bottom: 1px solid ${({ theme }) => theme.colors.gray100};
+  display: flex;
+  flex-direction: column;
+
+  &:last-child {
+    border-bottom: none;
+  }
 `;
 
-const StyledQnaCategory = styled.div<{ isActive: boolean }>`
-  display: flex;
-  align-items: center;
-  height: 48px;
-  padding: 0 10%;
-  position: relative;
-  gap: 5px;
-  cursor: pointer;
+const StyledQnaContentWrapper = styled.div`
+  height: 170px;
+  display: inline-flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 10px;
+`;
 
-  ::before {
-    content: '';
-    display: block;
-    height: 25px;
-    width: 3px;
-    position: absolute;
-    left: 0px;
-    border-left: 3px solid
-      ${({ theme, isActive }) => (isActive ? theme.colors.primary : 'transparent')};
-  }
+const StyledQnaContent = styled(Text)`
+  height: 46px;
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  -webkit-box-orient: vertical;
+  overflow: hidden;
 `;
