--- apps/client/src/store/user/qnaLengthAtom.ts
@@ -1,3 +1,0 @@
-import { atom } from 'jotai';
-
-export const qnaLengthAtom = atom(0);
